+++ spec/ci/node_15_spec.rb
@@ -1,9 +1,9 @@
 require_relative '../spec_helper'
 
-describe "Hello World for Node v11.x" do
-  context "a single-process Node v11.x app" do
+describe "Hello World for Node v15.x" do
+  context "a single-process Node v15.x app" do
     let(:app) {
-      Hatchet::Runner.new("spec/fixtures/repos/node-11")
+      Hatchet::Runner.new("spec/fixtures/repos/node-15")
     }
 
     it "should deploy successfully" do